requirements.txt
Flask==2.3.4
pandas==2.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
